backend/requirements.txt
Django>=5.1.4,<5.2
djangorestframework>=3.15.2,<3.16
requests>=2.31.0,<2.32
psycopg2>=2.9.7,<2.10
drf-spectacular>=0.26.3,<0.27
django-filter>=23.2,<23.3
Pillow>=10.0.0,<10.1
setuptools>=75.0.0,<76.0.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
urllib3>=2.2.0,<2.3.0
django-cors-headers>=4.5.0,<4.6.0
django-rest-passwordreset>=1.4.0,<1.5.0
openpyxl==3.1.5
gunicorn>=20.0.0,<21.0.0
